mega bunch of changes mostly to do with picking up project again and locationHandling

+++ wear/src/main/java/uk/me/ponies/wearroutes/WearRoutesApplication.java
@@ -1,22 +1,38 @@
 package uk.me.ponies.wearroutes;
 
 import android.app.Application;
+import android.content.Intent;
 import android.os.Environment;
+import android.util.Log;
 
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 
+import uk.me.ponies.wearroutes.locationService.LocationPollingService;
+import uk.me.ponies.wearroutes.utils.ActivityLifeCycleLogger;
+
 /**
- * Created by rummy on 02/09/2016.
+ * Used to provide an uncaught exception handler and other "global" like stuff
  */
 public class WearRoutesApplication extends Application {
-    public class MyApplication extends Application
-    {
-        public void onCreate ()
+        Intent locationServiceIntent;
+    private Thread.UncaughtExceptionHandler mDefaultUEHander;
+    private static final String TAG = "WearRoutesApplication";
+    ActivityLifeCycleLogger mActivityLifecycleLogger;
+
+    public void onCreate ()
         {
+            if (mActivityLifecycleLogger  == null) {
+                mActivityLifecycleLogger = new ActivityLifeCycleLogger();
+                registerActivityLifecycleCallbacks(mActivityLifecycleLogger);
+            } else {
+                Log.e(TAG, "onCreate called and lifecycle logger already registered!");
+            }
+
+
+            mDefaultUEHander = Thread.getDefaultUncaughtExceptionHandler();
             // Setup handler for uncaught exceptions.
             Thread.setDefaultUncaughtExceptionHandler (new Thread.UncaughtExceptionHandler()
             {
@@ -26,7 +42,23 @@ public class WearRoutesApplication extends Application {
                     handleUncaughtException (thread, e);
                 }
             });
+
+            locationServiceIntent = new Intent(this, LocationPollingService.class);
+            startService(locationServiceIntent);
             super.onCreate();
+        }
+
+        @Override
+        public void onTerminate() {
+            stopService(locationServiceIntent);
+            if (mActivityLifecycleLogger != null) {
+                unregisterActivityLifecycleCallbacks(mActivityLifecycleLogger);
+                mActivityLifecycleLogger = null;
+            }
+            else {
+                Log.e(TAG, "onTerminate called and no activityLifecycleLogger set");
+            }
+            super.onTerminate();
         }
 
         public void handleUncaughtException (Thread thread, Throwable e)
@@ -43,8 +75,12 @@ public class WearRoutesApplication extends Application {
             } catch (IOException ioe) {
                 // drop it .. we're dead anyway!
             }
-
-            System.exit(1); // kill off the crashed app
+            if (mDefaultUEHander != null) {
+                mDefaultUEHander.uncaughtException(thread, e);
+            }
+            else {
+                System.exit(1); // kill off the crashed app
+            }
         }
     }
-}
+
